Type Definition clamor_runtime::AllPalletsWithSystemReversed

pub type AllPalletsWithSystemReversed = (Accounts, Utility, Identity, Proxy, Multisig, Detach, Fragments, Protos, Contracts, Indices, Assets, Sudo, TransactionPayment, Balances, Grandpa, Aura, Timestamp, RandomnessCollectiveFlip, System);

👎Deprecated:
Using reverse pallet orders is deprecated. use only AllPalletWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order.